recommendationSystem/requirements.txt
pandas
requests
supabase
psycopg2-binary
python-dotenv
flask
